+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Features:
 - Network can set custom name and item icon now.
+- Added toast notification for order status change.
+- Ignore data ingredient is now allowed for batch crafting & large crafting, but each slot cannot have different data item.
+- Added "Custom Arrangement" mode in gauge; you can customize which item needs to be in which slot, and packager can only unpackage items to correct slot in target inventory.
 
 Changes:
 - Maximum batch crafting multiplier is 64 now, regardless recipe output count.